replacing authorization and authentication HOC with hooks
