+++ src/components/Session/context.js
@@ -3,10 +3,11 @@ import { useAuthentication } from ".";
 
 const AuthUserContext = React.createContext(null);
 
-export const AuthUserContextProvider = ({ children }) => (
+const AuthUserContextProvider = ({ children }) => (
   <AuthUserContext.Provider value={useAuthentication()}>
     {children}
   </AuthUserContext.Provider>
 );
 
 export default AuthUserContext;
+export { AuthUserContextProvider };
